fix: error in build
